refactor(unauth_respurce_section): move loginDialog to AddEvaluation
