+++ src/components/resource-section/add-evaluation/AddEvaluation.js
@@ -1,9 +1,12 @@
 import useCurrentUser from '@hooks/useCurrentUser';
+import useLoginDialog from '@hooks/useLoginDialog';
 import AddEvaluationAuth from './AddEvaluationAuth';
 import AddEvaluationAuthWidgets from './AddEvaluationAuthWidgets';
 
 const AddEvaluation = ({ resourceId, updatesAddEvaluation }) => {
   const currentUser = useCurrentUser();
+  const loginDialog = useLoginDialog();
+
   if (currentUser) {
     return (
       <AddEvaluationAuth
@@ -21,7 +24,21 @@ const AddEvaluation = ({ resourceId, updatesAddEvaluation }) => {
     enabled: false,
   };
 
-  return <AddEvaluationAuthWidgets formOptions={formOptions} />;
+  return (
+    <div
+      className="cardEvaluation"
+      onClick={() => {
+        loginDialog.setDisplayLoginDialog(true);
+      }}
+      onKeyPress={null}
+      role="button"
+      tabIndex="0"
+      data-pr-tooltip="Ingresa para poder evaluar un recurso"
+      data-pr-position="left"
+    >
+      <AddEvaluationAuthWidgets formOptions={formOptions} />;
+    </div>
+  );
 };
 
 export default AddEvaluation;
